add authorization to edit delete restaurants and itens

+++ project.py
@@ -3,7 +3,7 @@ from flask import Flask, render_template, request, redirect,jsonify, url_for, fl
 
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Restaurant, MenuItem
+from database_setup import Base, Restaurant, MenuItem, User
 
 from flask import session as login_session
 import random
@@ -24,7 +24,7 @@ APPLICATION_NAME = "Restaurant Menu Application"
 
 
 #Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenuoauth.db')
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -113,6 +113,13 @@ def gconnect():
   login_session['picture'] = data['picture']
   login_session['email'] = data['email']
 
+  user_id = getUserID(login_session['email'])
+
+  if user_id is None:
+    user_id = createUser(login_session)
+    
+  login_session['user_id'] = user_id
+  
   output = ''
   output += '<h1>Welcome, '
   output += login_session['username']
@@ -120,6 +127,7 @@ def gconnect():
   output += '<img src="'
   output += login_session['picture']
   output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+  
   flash("you are now logged in as %s" % login_session['username'])
   print "done!"
   return output
@@ -153,6 +161,28 @@ def gdisconnect():
     response.headers['Content-Type'] = 'application/json'
     return response
 
+# User Helper Functions
+
+def createUser(login_session):
+  newUser = User(name=login_session['username'], email=login_session['email'], 
+    picture=login_session['picture'])
+  session.add(newUser)
+  session.commit()
+  user = session.query(User).filter_by(email=login_session['email']).one()
+  return user.id
+
+
+def getUserInfo(user_id):
+  user = session.query(User).filter_by(id=user_id).one()
+  return user
+
+
+def getUserID(email):
+  try:
+    user = session.query(User).filter_by(email=email).one()
+    return user.id
+  except:
+    return None
 
 #JSON APIs to view Restaurant Information
 @app.route('/restaurant/<int:restaurant_id>/menu/JSON')
@@ -167,7 +197,7 @@ def menuItemJSON(restaurant_id, menu_id):
   Menu_Item = session.query(MenuItem).filter_by(id = menu_id).one()
   return jsonify(Menu_Item = Menu_Item.serialize)
 
-@app.route('/restaurant/JSON')
+@app.route('/restaurants/JSON')
 def restaurantsJSON():
   restaurants = session.query(Restaurant).all()
   return jsonify(restaurants= [r.serialize for r in restaurants])
@@ -175,10 +205,13 @@ def restaurantsJSON():
 
 #Show all restaurants
 @app.route('/')
-@app.route('/restaurant/')
+@app.route('/restaurants/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+  if 'username' not in login_session:
+    return render_template('publicrestaurants.html', restaurants=restaurants)
+  else:
+    return render_template('restaurants.html', restaurants=restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
@@ -188,7 +221,7 @@ def newRestaurant():
     return redirect('/login')
 
   if request.method == 'POST':
-    newRestaurant = Restaurant(name = request.form['name'])
+    newRestaurant = Restaurant(name = request.form['name'], user_id=login_session['user_id'])
     session.add(newRestaurant)
     flash('New Restaurant %s Successfully Created' % newRestaurant.name)
     session.commit()
@@ -204,6 +237,10 @@ def editRestaurant(restaurant_id):
     return redirect('/login')
 
   editedRestaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if editedRestaurant.user_id != login_session['user_id']:
+    return "<script>function myFunction() {alert('You are not authorized to edit this restaurant. Please create your own restaurant in order to edit.');}</script><body onload='myFunction()''>"
+
   if request.method == 'POST':
     if request.form['name']:
       editedRestaurant.name = request.form['name']
@@ -221,6 +258,10 @@ def deleteRestaurant(restaurant_id):
     return redirect('/login')
 
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if restaurantToDelete.user_id != login_session['user_id']:
+    return "<script>function myFunction() {alert('You are not authorized to delete this restaurant. Please create your own restaurant in order to delete.');}</script><body onload='myFunction()''>"
+
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -234,8 +275,13 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  creator = getUserInfo(restaurant.user_id)
   items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-  return render_template('menu.html', items = items, restaurant = restaurant)
+  if 'username' not in login_session or creator.id != login_session['user_id']:
+    return render_template('publicmenu.html', items=items, restaurant=restaurant, creator=creator)
+  else:
+    return render_template('menu.html', items=items, restaurant=restaurant, creator=creator)
+
      
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
@@ -245,8 +291,12 @@ def newMenuItem(restaurant_id):
     return redirect('/login')
 
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if login_session['user_id'] != restaurant.user_id:
+    return "<script>function myFunction() {alert('You are not authorized to add menu items to this restaurant. Please create your own restaurant in order to add items.');}</script><body onload='myFunction()'>"
+        
   if request.method == 'POST':
-    newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id)
+    newItem = MenuItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], course = request.form['course'], restaurant_id = restaurant_id, user_id=restaurant.user_id)
     session.add(newItem)
     session.commit()
     flash('New Menu %s Item Successfully Created' % (newItem.name))
@@ -263,6 +313,10 @@ def editMenuItem(restaurant_id, menu_id):
 
   editedItem = session.query(MenuItem).filter_by(id = menu_id).one()
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+
+  if login_session['user_id'] != restaurant.user_id:
+    return "<script>function myFunction() {alert('You are not authorized to edit menu items to this restaurant. Please create your own restaurant in order to edit items.');}</script><body onload='myFunction()'>"
+    
   if request.method == 'POST':
     if request.form['name']:
         editedItem.name = request.form['name']
@@ -289,6 +343,10 @@ def deleteMenuItem(restaurant_id,menu_id):
 
   restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
   itemToDelete = session.query(MenuItem).filter_by(id = menu_id).one() 
+
+  if login_session['user_id'] != restaurant.user_id:
+    return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()'>"
+    
   if request.method == 'POST':
     session.delete(itemToDelete)
     session.commit()
